Fix breaking change in rq-2.3.3
This fixes a breaking change in RQ where the internal location of
the `worker` method was moved from `rq.cli.cli.worker` to
`rq.cli.worker.worker`. The import was also available as
`rq.cli.worker` in past and present versions of RQ so we have changed
to referencing it from there.